added constructor for Item Class, loadU1 method, changed U1 class

+++ src/pl/juliuszmakowksi/U2.java
@@ -4,8 +4,8 @@ public class U2 extends Rocket{
     //field
 
     int cost = 120000000;
-    double rocketWeight = 18.0;
-    double maxWeightWithCargo = 29.0;
+    int rocketWeight = 18;
+    int maxWeightWithCargo = 29;
     double launchExplosionChance = 0.04;
     double landingExplosionChance = 0.08;
